refactor!: Major WIP revamp for abstraction and...
Mirror works differently than FishNet and made some assumptions. In addition to the revamp, we'll also get "stop" options (instead of just start). We'll also centralize all the button clicks instead of getting it from random the fishnet/mirror snippet helpers (that are incomplete). We also detach the UI serialization so we don't have to serialize it each time per-platform. Etc.