fix: Revert swapping back to old build (log loss)
The main point was it'd swap your platform back to your working one after you build a server. However, this causes an early recompile and build logs go away. I did keep the 100ms delay before the build (to let logs catch up) and a log to see the Hathora Server Config logs. This is also nice since you see a preview of the init build logs in hathora server config leading up to the build